Warning killer.
2001-04-30  Morten Welinder  <terra@diku.dk>

	* src/func.c (tokenized_help_find): Warning killer.

	* src/value.c (value_new_array): Kill warnings.
	(value_new_array_empty): Ditto.
	(value_release): Ditto.

	* src/print-cell.c (print_show): Use proper types.
	(get_width_string_n): Ditto.

	* src/number-match.c (gnumeric_error_init): Warning killer.

	* src/parser.y (deallocate_all): Warning killer.

	* src/item-bar.c (item_bar_draw): Avoid signedness problems.

	* src/format-template.c (hash_table_destroy): Warning killer.

	* src/colrow.h (struct _ColRowInfo): Make size_pixels an int.

	* src/value-sheet.c (value_area_get_width, value_area_get_height):
 	Return an int, not a guint.

2001-04-26  Morten Welinder  <terra@diku.dk>

	* fn-string.c (gnumeric_mid): Un-suckify.
	(gnumeric_find): Un-suckify.
	(gnumeric_substitute): Warning killer.
	(match_string): Warning killer.

	* fn-information.c (translate_cell_format): Warning killer.

	* fn-eng.c (val_to_base): Warning killer.

2001-04-30  Morten Welinder  <terra@diku.dk>

	* ms-summary.c (sum_name_to_excel): Fix index count.  Constify.
  	Make static.
	(excel_to_gnum_mapping): Make static.

+++ src/func.c
@@ -1029,11 +1029,11 @@ tokenized_help_find (TokenizedHelp *tok, const char *token)
 	if (!tok || !tok->sections)
 		return "Incorrect Function Description.";
 
-	for (lp = 0; lp + 1 < tok->sections->len; lp++) {
+	for (lp = 0; lp + 1 < (int)tok->sections->len; lp++) {
 		const char *cmp = g_ptr_array_index (tok->sections, lp);
 
 		if (g_strcasecmp (cmp, token) == 0){
-			return g_ptr_array_index (tok->sections, lp+1);
+			return g_ptr_array_index (tok->sections, lp + 1);
 		}
 	}
 	return "Cannot find token";
